refactor(#77): Resolve circular references
