+++ src/app/[lang]/adventures/[id]/AdventureDetailPage.page.tsx
@@ -2,8 +2,10 @@ import IdParamProps from "@app/_shared/idParam";
 import { render } from "@testing-library/react";
 import AdventureDetailPage from "./page";
 import { BasePage } from "@tests/pages/BasePage";
-import { mockAdventures } from "@features/adventures";
-import { StaticCardCarouselPageObject } from "@features/adventures/testExports";
+import {
+  StaticCardCarouselPageObject,
+} from "@features/adventures/testExports";
+import { mockAdventures } from "@tests/mockData/mockAdventures";
 
 export class AdventureDetailPageObject extends BasePage<IdParamProps> {
   private carouselPage!: StaticCardCarouselPageObject;
