+++ src/app/[lang]/adventures/[id]/AdventureDetailPage.spec.ts
@@ -1,6 +1,6 @@
 import { prismaMock } from "@tests/setup/prisma";
 import { AdventureDetailPageObject } from "./AdventureDetailPage.page";
-import { mockAdventures } from "@features/adventures";
+import { mockAdventures } from "@tests/mockData/mockAdventures";
 
 async function initPage(
   id: Number,
