+++ src/domain/models/adventure.ts
@@ -1,0 +1,3 @@
+import { Prisma } from "@prisma/client";
+
+export type Adventure = Prisma.AdventureGetPayload<{}>;
