+++ src/features/adventures/repositories/deleteAdventures.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { adventureRepository } from "../adventureRepository";
+import { adventureRepository } from "../repositoryExports";
 
 export async function deleteAdventureById(id: number) {
   return adventureRepository.deleteById(id);
