+++ src/features/adventures/repositoryExports.ts
@@ -1,4 +1,5 @@
 import prismaClient from "@repositories/prisma";
 import AdventureRepository from "./repositories/AdventureRepository";
 
+export { default as AdventureRepository } from "./repositories/AdventureRepository";
 export const adventureRepository = new AdventureRepository(prismaClient);
