+++ src/features/adventures/types/adventure.ts
@@ -1,3 +1,4 @@
+import { Adventure } from "@domain/models/adventure";
 import { Prisma } from "@prisma/client";
 
 export const adventureIncludes = {
@@ -6,8 +7,6 @@ export const adventureIncludes = {
   genres: true,
   series: true,
 };
-
-export type Adventure = Prisma.AdventureGetPayload<{}>;
 
 export type AdventureFull = Prisma.AdventureGetPayload<{
   include: { tags: true; system: true; genres: true; series: true };
